Nueva propuesta de Dominio de control.
