+++ es.uji.control.domain/src/es/uji/control/domain/ControlNotImplementedException.java
@@ -1,0 +1,11 @@
+package es.uji.control.domain;
+
+public class ControlNotImplementedException extends Exception {
+
+	private static final long serialVersionUID = -664857390469990142L;
+
+	public ControlNotImplementedException(String message) {
+		super(message);
+	}
+	
+}
