--- es.uji.control.domain/src/es/uji/control/domain/ControlServiceNotImplementedException.java
@@ -1,11 +1,0 @@
-package es.uji.control.domain;
-
-public class ControlServiceNotImplementedException extends Exception {
-
-	private static final long serialVersionUID = -664857390469990142L;
-
-	public ControlServiceNotImplementedException(String message) {
-		super(message);
-	}
-	
-}
