+++ es.uji.control.domain/src/es/uji/control/domain/IControlConnection.java
@@ -4,7 +4,7 @@ import es.uji.control.domain.authorizations.IAuthorizationsService;
 import es.uji.control.domain.people.IPersonService;
 
 public interface IControlConnection {
-	abstract IPersonService getPersonService() throws ControlServiceNotImplementedException;
-	abstract IAuthorizationsService getAuthorizationsService() throws ControlServiceNotImplementedException;
-	abstract void reset();
+	abstract IPersonService getPersonService() throws ControlNotImplementedException;
+	abstract IAuthorizationsService getAuthorizationsService() throws ControlNotImplementedException;
+	abstract void close();
 }
