+++ es.uji.control.domain/src/es/uji/control/domain/IControlConnectionFactorySelector.java
@@ -1,0 +1,6 @@
+package es.uji.control.domain;
+
+public interface IControlConnectionFactorySelector {
+	public abstract String getCurrentFactory();
+	public abstract void setCurrentFactory();
+}
