--- es.uji.control.domain/src/es/uji/control/domain/authorizations/IAuthorization.java
@@ -1,5 +1,0 @@
-package es.uji.control.domain.authorizations;
-
-public interface IAuthorization {
-
-}
